+++ operator/cmd/mesh/manifest-generate.go
@@ -37,11 +37,11 @@ import (
 )
 
 type manifestGenerateArgs struct {
-	// inFilenames is an array of paths to the input IstioOperator CR files.
+	// inFilenames is an array of paths to the input IstioControlPlane CR files.
 	inFilename []string
 	// outFilename is the path to the generated output directory.
 	outFilename string
-	// set is a string with element format "path=value" where path is an IstioOperator path and the value is a
+	// set is a string with element format "path=value" where path is an IstioControlPlane path and the value is a
 	// value to set the node at that path to.
 	set []string
 	// force proceeds even if there are validation errors
@@ -140,7 +140,7 @@ func GenManifests(inFilename []string, setOverlayYAML string, force bool,
 		return nil, nil, err
 	}
 
-	cp, err := controlplane.NewIstioOperator(mergedIOPS, t)
+	cp, err := controlplane.NewIstioControlPlane(mergedIOPS, t)
 	if err != nil {
 		return nil, nil, err
 	}
