+++ operator/cmd/mesh/operator-common.go
@@ -16,10 +16,7 @@ package mesh
 
 import (
 	"context"
-	"fmt"
 
-	v1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
 	v12 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 	_ "k8s.io/client-go/plugin/pkg/client/auth" // Import all Kubernetes client auth plugins (e.g. Azure, GCP, OIDC, etc.)
@@ -50,7 +47,7 @@ const (
 
 // isControllerInstalled reports whether an operator deployment exists in the given namespace.
 func isControllerInstalled(cs kubernetes.Interface, operatorNamespace string) (bool, error) {
-	return DeploymentExists(cs, operatorNamespace, operatorResourceName)
+	return deploymentExists(cs, operatorNamespace, operatorResourceName)
 }
 
 // renderOperatorManifest renders a manifest to install the operator with the given input arguments.
@@ -94,30 +91,8 @@ tag: {{.Tag}}
 	return vals, manifest, err
 }
 
-func CreateNamespace(cs kubernetes.Interface, namespace string) error {
-	if namespace == "" {
-		// Setup default namespace
-		namespace = istioDefaultNamespace
-	}
-
-	ns := &v1.Namespace{ObjectMeta: v12.ObjectMeta{
-		Name: namespace,
-		Labels: map[string]string{
-			"istio-injection": "disabled",
-		},
-	}}
-	_, err := cs.CoreV1().Namespaces().Create(context.TODO(), ns, v12.CreateOptions{})
-	if err != nil && !errors.IsAlreadyExists(err) {
-		return fmt.Errorf("failed to create namespace %v: %v", namespace, err)
-	}
-	return nil
-}
-
-func DeleteNamespace(cs kubernetes.Interface, namespace string) error {
-	return cs.CoreV1().Namespaces().Delete(context.TODO(), namespace, v12.DeleteOptions{})
-}
-
-func DeploymentExists(cs kubernetes.Interface, namespace, name string) (bool, error) {
+// deploymentExists returns true if the given deployment in the namespace exists.
+func deploymentExists(cs kubernetes.Interface, namespace, name string) (bool, error) {
 	d, err := cs.AppsV1().Deployments(namespace).Get(context.TODO(), name, v12.GetOptions{})
 	if err != nil {
 		return false, err
